Podcast: split distribution apps into self-contained registry
