--- projects/packages/podcast/src/dashboard/components/logos.tsx
@@ -1,156 +1,0 @@
-/**
- * Brand SVG logos for the podcast Distribution tab.
- *
- * Lifted unmodified from Calypso's `client/my-sites/podcast/components/logos.tsx`.
- * Each logo is rendered at 40×40 inside a flex row; alt text is suppressed via
- * aria-hidden because the directory name is rendered alongside.
- */
-
-import { useId } from '@wordpress/element';
-
-export const LogoApple = () => {
-	const gradientId = useId();
-	return (
-		<svg
-			xmlns="http://www.w3.org/2000/svg"
-			width="40"
-			height="40"
-			viewBox="0 0 300 300"
-			aria-hidden="true"
-			focusable="false"
-		>
-			<linearGradient
-				id={ gradientId }
-				gradientUnits="userSpaceOnUse"
-				x1="150"
-				y1="0"
-				x2="150"
-				y2="300"
-			>
-				<stop offset="0" style={ { stopColor: '#F452FF' } } />
-				<stop offset="1" style={ { stopColor: '#832BC1' } } />
-			</linearGradient>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				fill={ `url(#${ gradientId })` }
-				d="M294.1,260c-2.9,7.4-9.6,17.8-19.2,25.2c-5.5,4.2-12.1,8.3-21.1,11c-9.6,2.9-21.5,3.8-36.3,3.8h-135c-14.8,0-26.6-1-36.3-3.8c-9-2.7-15.6-6.7-21.1-11c-9.5-7.3-16.3-17.8-19.2-25.2C0.1,245.1,0,228.2,0,217.5l0,0v-135l0,0C0,71.8,0.1,54.9,5.9,40c2.9-7.4,9.6-17.8,19.2-25.2c5.5-4.2,12.1-8.3,21.1-11C55.9,1,67.7,0,82.5,0l0,0h135l0,0c14.8,0,26.6,1,36.3,3.8c9,2.7,15.6,6.7,21.1,11c9.5,7.3,16.3,17.8,19.2,25.2c5.9,14.9,5.9,31.9,5.9,42.5v135C300,228.2,299.9,245.1,294.1,260z"
-			/>
-			<path
-				fill="#FFFFFF"
-				d="M175.7,181.1c-0.4-3.6-1.6-6.2-4-8.6c-4.5-4.7-12.4-7.8-21.7-7.8c-9.3,0-17.2,3-21.7,7.8c-2.3,2.5-3.6,5-4,8.6c-0.8,7-0.3,13,0.5,22.7c0.8,9.2,2.3,21.5,4.2,33.9c1.4,8.9,2.5,13.7,3.5,17.1c1.7,5.6,7.8,10.4,17.5,10.4c9.7,0,15.9-4.9,17.5-10.4c1-3.4,2.1-8.2,3.5-17.1c1.9-12.5,3.4-24.7,4.2-33.9C176.1,194.1,176.5,188.1,175.7,181.1z"
-			/>
-			<path
-				fill="#FFFFFF"
-				d="M174.6,130.1c0,13.6-11,24.6-24.6,24.6s-24.6-11-24.6-24.6c0-13.6,11-24.6,24.6-24.6S174.6,116.6,174.6,130.1z"
-			/>
-			<path
-				fill="#FFFFFF"
-				d="M149.7,33.2C92.3,33.4,45.3,80,44.5,137.4c-0.6,46.5,29.1,86.3,70.6,100.9c1,0.4,2-0.5,1.9-1.5c-0.5-3.6-1.1-7.2-1.5-10.8c-0.2-1.3-1-2.3-2.1-2.9c-32.8-14.3-55.7-47.2-55.3-85.3c0.5-50,41.3-90.7,91.2-91.1c51.1-0.4,92.8,41,92.8,92c0,37.7-22.8,70.1-55.3,84.4c-1.2,0.5-2,1.6-2.1,2.9c-0.5,3.6-1,7.2-1.5,10.8c-0.2,1.1,0.9,1.9,1.9,1.5c41.1-14.4,70.6-53.6,70.6-99.6C255.5,80.5,208,33.1,149.7,33.2z"
-			/>
-			<path
-				fill="#FFFFFF"
-				d="M147.3,68.2c-37.4,1.4-67.4,32.3-67.9,69.7c-0.3,24.6,12,46.4,30.9,59.3c0.9,0.6,2.2-0.1,2.2-1.2c-0.3-4.3-0.3-8.1-0.1-12.1c0.1-1.3-0.4-2.5-1.4-3.4c-11.5-10.8-18.5-26.2-18.1-43.2c0.8-30,24.9-54.4,54.9-55.6c32.6-1.3,59.4,24.9,59.4,57.1c0,16.4-7,31.2-18.1,41.7c-0.9,0.9-1.4,2.1-1.4,3.4c0.2,3.9,0.1,7.7-0.1,12c-0.1,1.1,1.2,1.9,2.2,1.2c18.6-12.7,30.9-34.2,30.9-58.4C220.8,98.9,187.5,66.6,147.3,68.2z"
-			/>
-		</svg>
-	);
-};
-
-export const LogoSpotify = () => (
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		width="40"
-		height="40"
-		viewBox="0 0 220 220"
-		aria-hidden="true"
-		focusable="false"
-	>
-		<rect width="220" height="220" rx="44" ry="44" fill="#1ED760" />
-		<g transform="translate(26 26)">
-			<circle cx="83.996" cy="84.019" r="83.742" fill="#FFFFFF" />
-			<path
-				fill="#1ED760"
-				d="m122.4 120.78c-1.5 2.46-4.72 3.24-7.18 1.73-19.662-12.01-44.414-14.73-73.564-8.07-2.809 0.64-5.609-1.12-6.249-3.93-0.643-2.81 1.11-5.61 3.926-6.25 31.9-7.288 59.263-4.15 81.337 9.34 2.46 1.51 3.24 4.72 1.73 7.18zm10.25-22.802c-1.89 3.072-5.91 4.042-8.98 2.152-22.51-13.836-56.823-17.843-83.448-9.761-3.453 1.043-7.1-0.903-8.148-4.35-1.04-3.453 0.907-7.093 4.354-8.143 30.413-9.228 68.222-4.758 94.072 11.127 3.07 1.89 4.04 5.91 2.15 8.976v-0.001zm0.88-23.744c-26.99-16.031-71.52-17.505-97.289-9.684-4.138 1.255-8.514-1.081-9.768-5.219-1.254-4.14 1.08-8.513 5.221-9.771 29.581-8.98 78.756-7.245 109.83 11.202 3.73 2.209 4.95 7.016 2.74 10.733-2.2 3.722-7.02 4.949-10.73 2.739z"
-			/>
-		</g>
-	</svg>
-);
-
-export const LogoYouTube = () => (
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		width="40"
-		height="40"
-		viewBox="0 0 300 300"
-		aria-hidden="true"
-		focusable="false"
-	>
-		<rect width="300" height="300" fill="red" ry="49.137924" />
-		<path
-			fill="#fff"
-			d="M149.93750587 79.22267364s-63.21885.000059-79.08593926 4.12304707c-8.49578714 2.37380204-15.49140825 9.36950015-17.86523429 17.99023729-4.12295306 15.86710025-4.12304706 48.72656078-4.12304706 48.72656078s.000094 32.98420052 4.12304706 48.60156077c2.37382604 8.62062014 9.24450615 15.49138025 17.86523429 17.86524029 15.99203025 4.24788007 79.08593926 4.24804007 79.08593926 4.24804007s63.34418101-.00005 79.21094127-4.12304007c8.62079014-2.37381004 15.49133025-9.11966015 17.74023028-17.86524029 4.24793007-15.74232025 4.24805007-48.60156077 4.24805007-48.60156077s.12484-32.98446053-4.24805007-48.85156078c-2.24890003-8.62073714-9.11944014-15.49133425-17.74023028-17.74023729-15.86676026-4.37284707-79.21094127-4.37304707-79.21094127-4.37304707zm-20.11523032 40.48046465 52.59961084 30.35938049-52.59961084 30.23438048v-60.59376097z"
-		/>
-	</svg>
-);
-
-export const LogoAmazon = () => (
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		width="40"
-		height="40"
-		viewBox="0 0 1024 1024"
-		aria-hidden="true"
-		focusable="false"
-	>
-		<rect fill="#25d1da" width="1024" height="1024" rx="227" ry="227" />
-		<path
-			fill="#14191a"
-			d="M301.64,497.24c14.49,0,25.18,8.09,30.74,21.96h1.17c8.64-14.75,19.17-21.96,32.49-21.96,21.96,0,35.28,18.44,35.28,48.16v72.38c0,2.2-1.61,3.81-3.81,3.81h-23.42c-2.2,0-3.81-1.61-3.81-3.81v-70.04c0-14.2-5.56-22.1-15.95-22.1-11.42,0-17.42,9.22-17.42,26.2v65.94c0,2.2-1.61,3.81-3.81,3.81h-23.42c-2.2,0-3.81-1.61-3.81-3.81v-70.04c0-14.2-5.56-22.1-15.95-22.1-11.42,0-17.42,9.22-17.42,26.2v65.94c0,2.2-1.61,3.81-3.81,3.81h-23.42c-2.2,0-3.81-1.61-3.81-3.81v-114.1c0-2.2,1.61-3.81,3.81-3.81h19.32c2.2,0,3.62,1.46,4.1,3.81l2.93,14.34h1.17c5.27-13.32,15.37-20.78,28.84-20.78Z"
-		/>
-		<path
-			fill="#14191a"
-			d="M312.7,676.62c1.01,3.08,3.37,6.87,6.88,11.26,44.93,55.42,117.22,79.69,187.03,79.69,61.03,0,120-18.42,157.47-50.88,6.44-5.7,9.8-10.24,9.8-13.74,0-1.46-.59-2.63-1.61-3.8-3.22-3.36-8.05-.88-13.9,2.19-34.83,18.28-88.39,30.85-145.32,30.85-64.24,0-133.03-15.94-185.12-57.9-4.97-3.95-10.39-7.9-14.05-4.97-1.32,1.02-2.51,3.22-1.17,7.31Z"
-		/>
-		<path
-			fill="#14191a"
-			d="M637.43,682.76c1.47,3.36,5.71,3.36,12.29,2.05,17.56-3.8,39.8-4.1,43.61,1.17,3.66,5.41-4.39,26.17-13.9,41.38-3.51,5.85-4.98,9.5-2.05,12.28.73.88,1.76,1.17,2.93,1.32,3.51,0,8.05-3.51,12.29-7.46,9.22-8.04,16.39-18.28,20.34-29.83,3.95-11.11,5.85-26.76,1.76-32.75-3.95-5.7-19.32-9.5-31.17-9.36-12.44,0-24.44,3.07-35.27,8.77-6.73,3.66-12.87,7.75-10.83,12.43Z"
-		/>
-	</svg>
-);
-
-export const LogoPocketCasts = () => (
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		width="40"
-		height="40"
-		viewBox="0 0 32 32"
-		aria-hidden="true"
-		focusable="false"
-	>
-		<circle cx="16" cy="15" r="15" fill="white" />
-		<path
-			fillRule="evenodd"
-			clipRule="evenodd"
-			fill="#f43e37"
-			d="M16 32c8.837 0 16-7.163 16-16S24.837 0 16 0 0 7.163 0 16s7.163 16 16 16Zm0-28.444C9.127 3.556 3.556 9.127 3.556 16c0 6.873 5.571 12.444 12.444 12.444v-3.11A9.333 9.333 0 1 1 25.333 16h3.111c0-6.874-5.571-12.445-12.444-12.445ZM8.533 16A7.467 7.467 0 0 0 16 23.467v-2.715A4.751 4.751 0 1 1 20.752 16h2.715a7.467 7.467 0 0 0-14.934 0Z"
-		/>
-	</svg>
-);
-
-export const LogoPodcastIndex = () => (
-	<svg
-		xmlns="http://www.w3.org/2000/svg"
-		width="40"
-		height="40"
-		viewBox="0 0 60 61"
-		aria-hidden="true"
-		focusable="false"
-	>
-		<rect width="60" height="61" rx="12" ry="12" fill="#F90000" />
-		<g transform="translate(9.5 9.5)" fill="#FFFFFF">
-			<path d="M19.7935646,16 L21.1438236,16 C21.9722507,16 22.6438236,16.6715729 22.6438236,17.5 C22.6438236,17.6058285 22.632624,17.7113599 22.6104119,17.8148311 L17.4186047,42 L13,42 L18.3269763,17.1851689 C18.4753878,16.4938186 19.0864641,16 19.7935646,16 Z" />
-			<path d="M11.1084922,22.2287199 C6.355958,17.7559478 6.7537472,10.8190095 11.1294284,6.78109016 C11.9668794,6.01492086 13.3068009,6.05633542 14.0605068,6.90533384 C14.8560852,7.73362498 14.7932763,9.03818352 13.9558254,9.8250601 C11.2969186,12.2892262 11.0456833,16.4720965 13.9767617,19.2261645 C14.8142126,20.0130411 14.8560852,21.3175997 14.0605068,22.1458908 C13.2649284,22.974182 11.9459431,23.0155965 11.1084922,22.2287199 Z M9.64194995,25.2726899 C10.6887636,25.8939082 10.8981264,27.1363449 10.0606754,28.1717089 C9.64194995,28.7929272 9.01386174,29 8.38577352,29 C7.96704805,29 7.54832257,28.7929272 7.1295971,28.5858544 C2.94234234,25.4797627 0.220626747,20.0958702 0.0112640095,14.504905 C-0.198098728,8.70686707 2.52361686,3.94419301 7.1295971,0.423955663 C7.96704805,-0.197262692 9.43258721,-0.197262692 10.0606754,0.838101233 C10.6887636,1.66639237 10.6887636,3.11590187 9.64194995,3.73712022 C6.08278341,6.42906643 4.19851877,10.1563766 4.19851877,14.504905 C4.19851877,18.6463607 6.29214615,22.9948892 9.64194995,25.2726899 Z M26.9329847,22.1458908 C26.1374063,21.3175997 26.1792789,20.0130411 26.9957935,19.2054573 C29.9268719,16.4513892 29.6965729,12.268519 27.0167298,9.80435283 C26.1792789,9.03818352 26.11647,7.7129177 26.9120485,6.88462656 C27.6866906,6.05633542 29.0266121,5.99421358 29.8640631,6.78109016 C34.2397443,10.8190095 34.6375335,17.7559478 29.8849993,22.2287199 C29.0475484,23.0155965 27.7285631,22.974182 26.9329847,22.1458908 Z M33.872416,0.423955663 C38.4783962,3.94419301 41.1791755,8.70686707 40.9907491,14.504905 C40.9907491,20.0958702 38.2690335,25.4797627 33.872416,28.5858544 C33.4536905,28.7929272 33.034965,29 32.6162396,29 C31.9881514,29 31.3600631,28.7929272 30.9413377,28.1717089 C30.1038867,27.1363449 30.3132495,25.8939082 31.3600631,25.2726899 C34.7098669,22.9948892 36.8034943,18.6463607 36.8034943,14.504905 C36.8034943,10.1563766 34.9192297,6.42906643 31.3600631,3.73712022 C30.3132495,3.11590187 30.3132495,1.66639237 30.9413377,0.838101233 C31.5694259,-0.197262692 33.034965,-0.197262692 33.872416,0.423955663 Z" />
-			<circle cx="20.5" cy="14.5" r="4.5" />
-		</g>
-	</svg>
-);
