+++ projects/packages/podcast/src/dashboard/components/submit-modal.tsx
@@ -1,10 +1,13 @@
 /**
- * Three-step "submit your feed" modal launched from the Distribution tab.
+ * Default 3-step "submit your feed" modal launched from the Distribution tab.
  *
- * Ported from Calypso's `client/my-sites/podcast/components/submit-modal.tsx`,
- * minus the confetti animation. The submitted-show URL is stored on the
- * `podcasting_show_urls` site setting (validated server-side against a host
- * allowlist per directory) rather than per-browser localStorage.
+ * Conforms to `PodcastAppModalProps` so an app's `Modal` field can swap in a
+ * custom flow without distribution.tsx caring. Apps that fit this pattern
+ * leave `Modal` unset and ride this default; ones with diverging flows (e.g.
+ * one-click API submission) ship their own component instead.
+ *
+ * The submitted-show URL is persisted on the `podcasting_show_urls` site
+ * setting (server-side host allowlist enforced by class-settings-rest.php).
  */
 
 import {
@@ -24,27 +27,14 @@ import { useCallback, useEffect, useState } from '@wordpress/element';
 import { __, sprintf } from '@wordpress/i18n';
 import { external, link } from '@wordpress/icons';
 import { usePodcastSettings, useUpdatePodcastSettings } from '../hooks/use-podcast-settings';
-import type { PodcatcherId } from '../types';
+import type { PodcastAppModalProps } from '../podcast-apps';
 import type { FormEvent } from 'react';
 
-export interface Podcatcher {
-	id: PodcatcherId;
-	name: string;
-	submitUrl: string;
-	learnMoreUrl?: string;
-}
-
-interface SubmitModalProps {
-	feedUrl: string;
-	podcatcher: Podcatcher;
-	onClose: () => void;
-}
-
-const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
+const SubmitModal = ( { app, feedUrl, onClose }: PodcastAppModalProps ) => {
 	const { data: settings } = usePodcastSettings();
 	const { mutate: saveSettings, isPending: isSaving } = useUpdatePodcastSettings();
 
-	const storedUrl = settings?.podcasting_show_urls?.[ podcatcher.id ] ?? '';
+	const storedUrl = settings?.podcasting_show_urls?.[ app.id ] ?? '';
 	const [ draftUrl, setDraftUrl ] = useState( storedUrl );
 	const [ hasCopied, setHasCopied ] = useState( false );
 
@@ -62,11 +52,11 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 			event.preventDefault();
 			// Send only the changed key — server merges with the stored map.
 			saveSettings(
-				{ podcasting_show_urls: { [ podcatcher.id ]: draftUrl.trim() } },
+				{ podcasting_show_urls: { [ app.id ]: draftUrl.trim() } },
 				{ onSuccess: onClose }
 			);
 		},
-		[ draftUrl, podcatcher.id, saveSettings, onClose ]
+		[ draftUrl, app.id, saveSettings, onClose ]
 	);
 
 	const isUnchanged = draftUrl.trim() === storedUrl.trim();
@@ -79,13 +69,8 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 	const titleText = sprintf(
 		/* translators: %s: podcast directory name (e.g. "Apple Podcasts"). */
 		__( 'Submit to %s', 'jetpack-podcast' ),
-		podcatcher.name
+		app.name
 	);
-
-	const step2Note =
-		podcatcher.id === 'pocketcasts'
-			? __( 'Choose the Public option, since this feed is for your listeners.', 'jetpack-podcast' )
-			: null;
 
 	return (
 		<Modal title={ titleText } onRequestClose={ onClose } className="podcast__submit-modal">
@@ -102,7 +87,7 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 										'Click the button below to copy your RSS feed URL. %s will require this URL to list your podcast.',
 										'jetpack-podcast'
 									),
-									podcatcher.name
+									app.name
 							  )
 							: __(
 									'Set your podcast category in the Settings tab to generate your RSS feed URL.',
@@ -128,7 +113,7 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 						{ sprintf(
 							/* translators: %s: podcast directory name. */
 							__( 'Step 2: Submit your podcast to %s', 'jetpack-podcast' ),
-							podcatcher.name
+							app.name
 						) }
 					</h2>
 					<Text as="p" variant="muted">
@@ -138,39 +123,35 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 								'Click the button below to visit %s and complete their sign up flow.',
 								'jetpack-podcast'
 							),
-							podcatcher.name
+							app.name
 						) }
 					</Text>
-					{ podcatcher.learnMoreUrl && (
+					{ app.learnMoreUrl && (
 						<Text as="p" variant="muted">
-							<ExternalLink href={ podcatcher.learnMoreUrl }>
+							<ExternalLink href={ app.learnMoreUrl }>
 								{ __( 'Learn more', 'jetpack-podcast' ) }
 							</ExternalLink>
 						</Text>
 					) }
-					{ step2Note && (
-						<Text as="p" variant="muted">
-							{ step2Note }
-						</Text>
-					) }
+					{ app.step2Extra }
 					<Button
 						variant="secondary"
 						__next40pxDefaultSize
 						icon={ external }
 						iconPosition="right"
-						href={ podcatcher.submitUrl }
+						href={ app.submitUrl }
 						target="_blank"
 						rel="noopener noreferrer"
 						aria-label={ sprintf(
 							/* translators: %s: podcast directory name. */
 							__( 'Visit %s (opens in a new tab)', 'jetpack-podcast' ),
-							podcatcher.name
+							app.name
 						) }
 					>
 						{ sprintf(
 							/* translators: %s: podcast directory name. */
 							__( 'Visit %s', 'jetpack-podcast' ),
-							podcatcher.name
+							app.name
 						) }
 					</Button>
 				</VStack>
@@ -180,7 +161,7 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 						{ sprintf(
 							/* translators: %s: podcast directory name. */
 							__( 'Step 3: Enter your %s URL', 'jetpack-podcast' ),
-							podcatcher.name
+							app.name
 						) }
 					</h2>
 					<Text as="p" variant="muted">
@@ -190,7 +171,7 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 								'Paste your new %s URL into the field below and we’ll use it for your sharing buttons.',
 								'jetpack-podcast'
 							),
-							podcatcher.name
+							app.name
 						) }
 					</Text>
 					<form className="podcast__submit-step-form" onSubmit={ handleSave }>
@@ -200,7 +181,7 @@ const SubmitModal = ( { feedUrl, podcatcher, onClose }: SubmitModalProps ) => {
 									label={ sprintf(
 										/* translators: %s: podcast directory name. */
 										__( '%s URL', 'jetpack-podcast' ),
-										podcatcher.name
+										app.name
 									) }
 									hideLabelFromVision
 									value={ draftUrl }
