+++ projects/packages/podcast/src/dashboard/podcast-apps/amazon.tsx
@@ -1,0 +1,37 @@
+/**
+ * Amazon Music directory entry.
+ */
+
+import type { PodcastApp } from './types';
+
+const AmazonLogo = () => (
+	<svg
+		xmlns="http://www.w3.org/2000/svg"
+		width="40"
+		height="40"
+		viewBox="0 0 1024 1024"
+		aria-hidden="true"
+		focusable="false"
+	>
+		<rect fill="#25d1da" width="1024" height="1024" rx="227" ry="227" />
+		<path
+			fill="#14191a"
+			d="M301.64,497.24c14.49,0,25.18,8.09,30.74,21.96h1.17c8.64-14.75,19.17-21.96,32.49-21.96,21.96,0,35.28,18.44,35.28,48.16v72.38c0,2.2-1.61,3.81-3.81,3.81h-23.42c-2.2,0-3.81-1.61-3.81-3.81v-70.04c0-14.2-5.56-22.1-15.95-22.1-11.42,0-17.42,9.22-17.42,26.2v65.94c0,2.2-1.61,3.81-3.81,3.81h-23.42c-2.2,0-3.81-1.61-3.81-3.81v-70.04c0-14.2-5.56-22.1-15.95-22.1-11.42,0-17.42,9.22-17.42,26.2v65.94c0,2.2-1.61,3.81-3.81,3.81h-23.42c-2.2,0-3.81-1.61-3.81-3.81v-114.1c0-2.2,1.61-3.81,3.81-3.81h19.32c2.2,0,3.62,1.46,4.1,3.81l2.93,14.34h1.17c5.27-13.32,15.37-20.78,28.84-20.78Z"
+		/>
+		<path
+			fill="#14191a"
+			d="M312.7,676.62c1.01,3.08,3.37,6.87,6.88,11.26,44.93,55.42,117.22,79.69,187.03,79.69,61.03,0,120-18.42,157.47-50.88,6.44-5.7,9.8-10.24,9.8-13.74,0-1.46-.59-2.63-1.61-3.8-3.22-3.36-8.05-.88-13.9,2.19-34.83,18.28-88.39,30.85-145.32,30.85-64.24,0-133.03-15.94-185.12-57.9-4.97-3.95-10.39-7.9-14.05-4.97-1.32,1.02-2.51,3.22-1.17,7.31Z"
+		/>
+		<path
+			fill="#14191a"
+			d="M637.43,682.76c1.47,3.36,5.71,3.36,12.29,2.05,17.56-3.8,39.8-4.1,43.61,1.17,3.66,5.41-4.39,26.17-13.9,41.38-3.51,5.85-4.98,9.5-2.05,12.28.73.88,1.76,1.17,2.93,1.32,3.51,0,8.05-3.51,12.29-7.46,9.22-8.04,16.39-18.28,20.34-29.83,3.95-11.11,5.85-26.76,1.76-32.75-3.95-5.7-19.32-9.5-31.17-9.36-12.44,0-24.44,3.07-35.27,8.77-6.73,3.66-12.87,7.75-10.83,12.43Z"
+		/>
+	</svg>
+);
+
+export const amazon: PodcastApp = {
+	id: 'amazon',
+	name: 'Amazon Music',
+	Logo: AmazonLogo,
+	submitUrl: 'https://podcasters.amazon.com',
+};
