+++ projects/packages/podcast/src/dashboard/podcast-apps/apple.tsx
@@ -1,0 +1,62 @@
+/**
+ * Apple Podcasts directory entry.
+ */
+
+import { useId } from '@wordpress/element';
+import type { PodcastApp } from './types';
+
+const AppleLogo = () => {
+	const gradientId = useId();
+	return (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="40"
+			height="40"
+			viewBox="0 0 300 300"
+			aria-hidden="true"
+			focusable="false"
+		>
+			<linearGradient
+				id={ gradientId }
+				gradientUnits="userSpaceOnUse"
+				x1="150"
+				y1="0"
+				x2="150"
+				y2="300"
+			>
+				<stop offset="0" style={ { stopColor: '#F452FF' } } />
+				<stop offset="1" style={ { stopColor: '#832BC1' } } />
+			</linearGradient>
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				fill={ `url(#${ gradientId })` }
+				d="M294.1,260c-2.9,7.4-9.6,17.8-19.2,25.2c-5.5,4.2-12.1,8.3-21.1,11c-9.6,2.9-21.5,3.8-36.3,3.8h-135c-14.8,0-26.6-1-36.3-3.8c-9-2.7-15.6-6.7-21.1-11c-9.5-7.3-16.3-17.8-19.2-25.2C0.1,245.1,0,228.2,0,217.5l0,0v-135l0,0C0,71.8,0.1,54.9,5.9,40c2.9-7.4,9.6-17.8,19.2-25.2c5.5-4.2,12.1-8.3,21.1-11C55.9,1,67.7,0,82.5,0l0,0h135l0,0c14.8,0,26.6,1,36.3,3.8c9,2.7,15.6,6.7,21.1,11c9.5,7.3,16.3,17.8,19.2,25.2c5.9,14.9,5.9,31.9,5.9,42.5v135C300,228.2,299.9,245.1,294.1,260z"
+			/>
+			<path
+				fill="#FFFFFF"
+				d="M175.7,181.1c-0.4-3.6-1.6-6.2-4-8.6c-4.5-4.7-12.4-7.8-21.7-7.8c-9.3,0-17.2,3-21.7,7.8c-2.3,2.5-3.6,5-4,8.6c-0.8,7-0.3,13,0.5,22.7c0.8,9.2,2.3,21.5,4.2,33.9c1.4,8.9,2.5,13.7,3.5,17.1c1.7,5.6,7.8,10.4,17.5,10.4c9.7,0,15.9-4.9,17.5-10.4c1-3.4,2.1-8.2,3.5-17.1c1.9-12.5,3.4-24.7,4.2-33.9C176.1,194.1,176.5,188.1,175.7,181.1z"
+			/>
+			<path
+				fill="#FFFFFF"
+				d="M174.6,130.1c0,13.6-11,24.6-24.6,24.6s-24.6-11-24.6-24.6c0-13.6,11-24.6,24.6-24.6S174.6,116.6,174.6,130.1z"
+			/>
+			<path
+				fill="#FFFFFF"
+				d="M149.7,33.2C92.3,33.4,45.3,80,44.5,137.4c-0.6,46.5,29.1,86.3,70.6,100.9c1,0.4,2-0.5,1.9-1.5c-0.5-3.6-1.1-7.2-1.5-10.8c-0.2-1.3-1-2.3-2.1-2.9c-32.8-14.3-55.7-47.2-55.3-85.3c0.5-50,41.3-90.7,91.2-91.1c51.1-0.4,92.8,41,92.8,92c0,37.7-22.8,70.1-55.3,84.4c-1.2,0.5-2,1.6-2.1,2.9c-0.5,3.6-1,7.2-1.5,10.8c-0.2,1.1,0.9,1.9,1.9,1.5c41.1-14.4,70.6-53.6,70.6-99.6C255.5,80.5,208,33.1,149.7,33.2z"
+			/>
+			<path
+				fill="#FFFFFF"
+				d="M147.3,68.2c-37.4,1.4-67.4,32.3-67.9,69.7c-0.3,24.6,12,46.4,30.9,59.3c0.9,0.6,2.2-0.1,2.2-1.2c-0.3-4.3-0.3-8.1-0.1-12.1c0.1-1.3-0.4-2.5-1.4-3.4c-11.5-10.8-18.5-26.2-18.1-43.2c0.8-30,24.9-54.4,54.9-55.6c32.6-1.3,59.4,24.9,59.4,57.1c0,16.4-7,31.2-18.1,41.7c-0.9,0.9-1.4,2.1-1.4,3.4c0.2,3.9,0.1,7.7-0.1,12c-0.1,1.1,1.2,1.9,2.2,1.2c18.6-12.7,30.9-34.2,30.9-58.4C220.8,98.9,187.5,66.6,147.3,68.2z"
+			/>
+		</svg>
+	);
+};
+
+export const apple: PodcastApp = {
+	id: 'apple',
+	name: 'Apple Podcasts',
+	Logo: AppleLogo,
+	submitUrl: 'https://podcastsconnect.apple.com/',
+	learnMoreUrl: 'https://podcasters.apple.com/support/897-submit-a-show',
+};
