+++ projects/packages/podcast/src/dashboard/podcast-apps/podcastindex.tsx
@@ -1,0 +1,30 @@
+/**
+ * Podcast Index directory entry.
+ */
+
+import type { PodcastApp } from './types';
+
+const PodcastIndexLogo = () => (
+	<svg
+		xmlns="http://www.w3.org/2000/svg"
+		width="40"
+		height="40"
+		viewBox="0 0 60 61"
+		aria-hidden="true"
+		focusable="false"
+	>
+		<rect width="60" height="61" rx="12" ry="12" fill="#F90000" />
+		<g transform="translate(9.5 9.5)" fill="#FFFFFF">
+			<path d="M19.7935646,16 L21.1438236,16 C21.9722507,16 22.6438236,16.6715729 22.6438236,17.5 C22.6438236,17.6058285 22.632624,17.7113599 22.6104119,17.8148311 L17.4186047,42 L13,42 L18.3269763,17.1851689 C18.4753878,16.4938186 19.0864641,16 19.7935646,16 Z" />
+			<path d="M11.1084922,22.2287199 C6.355958,17.7559478 6.7537472,10.8190095 11.1294284,6.78109016 C11.9668794,6.01492086 13.3068009,6.05633542 14.0605068,6.90533384 C14.8560852,7.73362498 14.7932763,9.03818352 13.9558254,9.8250601 C11.2969186,12.2892262 11.0456833,16.4720965 13.9767617,19.2261645 C14.8142126,20.0130411 14.8560852,21.3175997 14.0605068,22.1458908 C13.2649284,22.974182 11.9459431,23.0155965 11.1084922,22.2287199 Z M9.64194995,25.2726899 C10.6887636,25.8939082 10.8981264,27.1363449 10.0606754,28.1717089 C9.64194995,28.7929272 9.01386174,29 8.38577352,29 C7.96704805,29 7.54832257,28.7929272 7.1295971,28.5858544 C2.94234234,25.4797627 0.220626747,20.0958702 0.0112640095,14.504905 C-0.198098728,8.70686707 2.52361686,3.94419301 7.1295971,0.423955663 C7.96704805,-0.197262692 9.43258721,-0.197262692 10.0606754,0.838101233 C10.6887636,1.66639237 10.6887636,3.11590187 9.64194995,3.73712022 C6.08278341,6.42906643 4.19851877,10.1563766 4.19851877,14.504905 C4.19851877,18.6463607 6.29214615,22.9948892 9.64194995,25.2726899 Z M26.9329847,22.1458908 C26.1374063,21.3175997 26.1792789,20.0130411 26.9957935,19.2054573 C29.9268719,16.4513892 29.6965729,12.268519 27.0167298,9.80435283 C26.1792789,9.03818352 26.11647,7.7129177 26.9120485,6.88462656 C27.6866906,6.05633542 29.0266121,5.99421358 29.8640631,6.78109016 C34.2397443,10.8190095 34.6375335,17.7559478 29.8849993,22.2287199 C29.0475484,23.0155965 27.7285631,22.974182 26.9329847,22.1458908 Z M33.872416,0.423955663 C38.4783962,3.94419301 41.1791755,8.70686707 40.9907491,14.504905 C40.9907491,20.0958702 38.2690335,25.4797627 33.872416,28.5858544 C33.4536905,28.7929272 33.034965,29 32.6162396,29 C31.9881514,29 31.3600631,28.7929272 30.9413377,28.1717089 C30.1038867,27.1363449 30.3132495,25.8939082 31.3600631,25.2726899 C34.7098669,22.9948892 36.8034943,18.6463607 36.8034943,14.504905 C36.8034943,10.1563766 34.9192297,6.42906643 31.3600631,3.73712022 C30.3132495,3.11590187 30.3132495,1.66639237 30.9413377,0.838101233 C31.5694259,-0.197262692 33.034965,-0.197262692 33.872416,0.423955663 Z" />
+			<circle cx="20.5" cy="14.5" r="4.5" />
+		</g>
+	</svg>
+);
+
+export const podcastindex: PodcastApp = {
+	id: 'podcastindex',
+	name: 'Podcast Index',
+	Logo: PodcastIndexLogo,
+	submitUrl: 'https://podcastindex.org/add',
+};
